* DEL: Removed duplicated code
* FIX: Fixed numeric types and boolean types
* FIX: Correctly pragma warning using
* FIX: Correction of the use of the purification operator.
* DEL: Removed dead and unused code
* ADD: Added ignoring of specific code into PVS_STUDIO
* And anothers minor fixes...

+++ code/engine.vc2008/xrCore/xrMemory.cpp
@@ -30,7 +30,7 @@ void xrMemory::_destroy()
 	mem_initialized = false;
 }
 
-inline const size_t external_size = -1;
+inline const size_t external_size = size_t(-1);
 
 void xrMemory::mem_compact()
 {
